requirements.txt
flask
pandas
gunicorn
werkzeug
openpyxl
psycopg2-binary

